Обновил библиотеку, написал новый генератор MyGson2
